feat(transaction): add store POS endpoint
Added a POST endpoint at /store/{idLocal}/transactions/{idPos} in
`TransactionController` to allow registering a new transaction
associated with a specific store and point of sale (POS).

Implemented the core logic in the `createTransaction` method of
`TransactionService`, which:
- Maps transaction details from the request DTO.
- Calculates each product's subtotal based on quantity and price.
- Associates the transaction with the active POS session.
- Calculates the total transaction amount.
- Persists the entity and returns the corresponding DTO.

New endpoint:
- POST /store/{idLocal}/transactions/{idPos}

Technical changes:
- Extended `TransactionService` constructor to inject additional
  repositories and mappers.
- Added logic for detail mapping, subtotal calculation, and total
  aggregation.

+++ src/main/java/org/stockify/dto/request/DetailTransactionRequest.java
@@ -5,6 +5,7 @@ import lombok.Setter;
 import lombok.Value;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 /**
  * DTO for {@link org.stockify.model.entity.DetailTransactionEntity}
@@ -14,5 +15,5 @@ import java.io.Serializable;
 @Setter
 public class DetailTransactionRequest{
     String productBarcode;
-    Long quantity;
+    BigDecimal quantity;
 }
